#281 Utilize unsigned type

+++ examples/simulate_fmi1_cs.c
@@ -26,7 +26,7 @@ int main(int argc, char* argv[]) {
     // initialize the FMU
     CALL(FMI1InitializeSlave(S, startTime, fmi1True, stopTime));
 
-    for (int step = 0;; step++) {
+    for (size_t step = 0;; step++) {
 
         CALL(recordVariables(S, outputFile));
